backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.6
celery>=5.3.1
redis>=4.6.0
django-cors-headers>=4.2.0
meilisearch>=0.28.1
requests>=2.31.0
paramiko>=3.3.1
dj-database-url>=2.1.0
python-dotenv>=1.0.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0